requirements.txt
requests>=2.28
click>=8.0
python-dotenv>=1.0
